+++ packages/dash-table/src/dash-table/components/Table/props.ts
@@ -319,6 +319,7 @@ interface IDefaultProps {
     fill_width: boolean;
     filter_query: string;
     filter_action: TableAction;
+    include_headers_on_copy_paste: boolean;
     merge_duplicate_headers: boolean;
     fixed_columns: Fixed;
     fixed_rows: Fixed;