+++ packages/dash-table/CHANGELOG.md
@@ -25,6 +25,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     row of `data`.
     - Additionally clearing the column will reset the filter for the affected column(s)
 
+[#318](https://github.com/plotly/dash-table/issues/318)
+- Headers are included when copying from the table to different 
+tabs and elsewhere. They are ignored when copying from the table onto itself and 
+between two tables within the same tab. 
+
 ### Changed
 [#497](https://github.com/plotly/dash-table/pull/497)
 - Like for clearing above, deleting through the `x` action will also